Move the getUploadsConfig into its own file
